+++ packages/core/src/lib/user-context/provide-user-context.ts
@@ -1,0 +1,6 @@
+import { makeEnvironmentProviders, EnvironmentProviders } from '@angular/core';
+import { UserContextService } from './user-context.service';
+
+export function provideUserContext(): EnvironmentProviders {
+  return makeEnvironmentProviders([UserContextService]);
+}